[TASK] Make job:work more resilient
The job:work command is made more resilient against errors during job
execution with this change.

Before, an exception thrown during job execution would lead to the
termination of the job:work command. Now a message is logged to the
standard output and the loop continues to run.

Also jobs are marked finished now in any case when waitAndExecute() on
the JobManager returns. Otherwise pending jobs would pile up that could
not be restarted anyway (since they were gone from the queue already).