!!! FEATURE: Allow defer annotation to specify custom options
This change adds an optional `option` argument to the `Defer`
annotation that allows for specifying arbitrary parameters that are
passed to the Queue when a job is queued::

  @Job\Defer(queueName="some-queue", options="{delay: 123}")
  public function deferredMethod() { //...

Note: The supported options depend on the specific queue
implementation. The `Flowpack.JobQueue.Beanstalkd` queue for example
supports the options "delay", "priority" and "ttr".

This is a breaking change because it changes the signature of
`QueueInterface::submit()` and `JobManager::queue()`.