TASK: Allow nested defer annotations
This will allow defer annotations to trigger creation of another job while executing a job. Previously this was completely prevented to avoid an infinite loop as the aspect is triggered within a job execution even for the method that the job tries to execute.
This method hash ensures that the same method cannot be deferred but any other method with a defer annotation can again become deferred.